# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Luna DevOps API
  version: 0.1.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/projects/{projectId}/applications:
    get:
      tags: [Applications]
      summary: List applications
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Application list.
    post:
      tags: [Applications]
      summary: Create application
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created application.
  /api/v1/projects/{projectId}/applications/{applicationId}:
    get:
      tags: [Applications]
      summary: Get application
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ApplicationId"
      responses:
        "200":
          description: Application.
    put:
      tags: [Applications]
      summary: Update application
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ApplicationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "200":
          description: Updated application.
    delete:
      tags: [Applications]
      summary: Delete application
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ApplicationId"
      responses:
        "204":
          description: Deleted application.
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: lyd_session
      description: HttpOnly browser session cookie. Sensitive interactive operations do not accept personal access tokens as a substitute.
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: Luna personal access token
  parameters:
...
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema: { type: string }
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    ApplicationInput:
      type: object
      properties:
        name: { type: string }
        identifier:
          type: string
          minLength: 2
          maxLength: 22
          pattern: "^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$"
          description: Immutable application identifier, unique within its project space and used to derive stable resource IDs.
        sourceType: { type: string, enum: [repository, image] }
        repositoryUrl: { type: string }
        imageReference: { type: string }
        dockerfilePath: { type: string }
        buildContext: { type: string }
        servicePort: { type: integer }
      required: [name, identifier, sourceType]
...